게임화 주요 업적 엔드포인트 도입, Dockerfile 및 GitHub Actions 배포 구성 추가, 그리고 인증·스토어·커리어 관련 DTO와 서비스 업데이트.

+++ gamification/entities/stamp.go
@@ -5,8 +5,8 @@ import (
 )
 
 type Stamp struct {
-	ID        uint    `gorm:"primaryKey;autoIncrement" json:"id"`
-	StoreID   *uint   `gorm:"column:storeId" json:"storeId"`
+	ID        int64   `gorm:"primaryKey;autoIncrement" json:"id"`
+	StoreID   *int64  `gorm:"column:storeId" json:"storeId"`
 	ImageName *string `gorm:"column:imageName;type:varchar(255)" json:"imageName"` // 가게 로고 혹은 지역 심볼
 
 	// Associations
